- Bug Squatting

+++ app/src/main/java/com/blackboxindia/TakeIT/Fragments/frag_ViewMyAd.java
@@ -13,12 +13,12 @@ import android.support.v7.widget.LinearLayoutManager;
 import android.support.v7.widget.RecyclerView;
 import android.util.Log;
 import android.view.LayoutInflater;
+import android.view.MenuItem;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.TextView;
 import android.widget.Toast;
 
-import com.blackboxindia.TakeIT.Network.CloudStorageMethods;
 import com.blackboxindia.TakeIT.Network.Interfaces.onDeleteListener;
 import com.blackboxindia.TakeIT.Network.NetworkMethods;
 import com.blackboxindia.TakeIT.R;
@@ -28,7 +28,6 @@ import com.blackboxindia.TakeIT.dataModels.AdData;
 import com.blackboxindia.TakeIT.dataModels.UserInfo;
 import com.google.firebase.auth.FirebaseAuth;
 
-import static com.blackboxindia.TakeIT.activities.MainActivity.MAIN_FRAG_TAG;
 import static com.blackboxindia.TakeIT.activities.MainActivity.MY_ADS_TAG;
 import static com.blackboxindia.TakeIT.activities.MainActivity.VIEW_MyAD_TAG;
 
@@ -41,61 +40,21 @@ public class frag_ViewMyAd extends Fragment {
     View view;
     Context context;
 
-
     AdData adData;
     Bitmap main;
-    CloudStorageMethods cloudStorageMethods;
     //endregion
 
     //region Initial Setup
-    @Nullable
     @Override
-    public View onCreateView(LayoutInflater inflater, @Nullable ViewGroup container, Bundle savedInstanceState) {
+    public void onResume() {
+        ((MainActivity)getActivity()).hideIT();
+        MenuItem item = ((MainActivity) getActivity()).toolbar.getMenu().findItem(R.id.toolbar_delete);
+        item.setVisible(true);
+        item.setOnMenuItemClickListener(new MenuItem.OnMenuItemClickListener() {
+            @Override
+            public boolean onMenuItemClick(MenuItem item) {
+                if(item.getItemId() == R.id.toolbar_delete){
 
-        view = inflater.inflate(R.layout.frag_viewmyad, container, false);
-        Log.i(TAG,"onCreateView");
-        context = view.getContext();
-
-        initVariables();
-
-        setUpViews();
-        return view;
-    }
-
-    private void initVariables() {
-
-        tv_Title = (TextView) view.findViewById(R.id.Ad_tvTitle);
-        tv_Price = (TextView) view.findViewById(R.id.Ad_tvPrice);
-        tv_Description = (TextView) view.findViewById(R.id.Ad_tvDescription);
-        imgRecyclerView = (RecyclerView) view.findViewById(R.id.Ad_imgRecycler);
-
-        cloudStorageMethods = new CloudStorageMethods(context);
-    }
-
-    void setUpViews() {
-
-        adData = getArguments().getParcelable("adData");
-
-        if(adData!=null) {
-            Log.i(TAG,"AdData not null");
-
-            if (adData.getPrice() == 0)
-                tv_Price.setText(getString(R.string.free));
-            else
-                tv_Price.setText(String.format(getString(R.string.currency), adData.getPrice()));
-
-            tv_Title.setText(adData.getTitle());
-            tv_Description.setText(adData.getDescription());
-
-            setUpImgRecycler();
-
-
-            ((MainActivity)context).ic_Toolbar.setVisibility(View.VISIBLE);
-            ((MainActivity)context).ic_Toolbar.setImageResource(R.drawable.ic_refresh);
-            ((MainActivity)context).ic_Toolbar.setImageResource(R.drawable.ic_delete);
-            ((MainActivity)context).ic_Toolbar.setOnClickListener(new View.OnClickListener() {
-                @Override
-                public void onClick(View v) {
                     final ProgressDialog dialog = ProgressDialog.show(context, "Deleting...", "", true, false);
                     NetworkMethods methods = new NetworkMethods(context, FirebaseAuth.getInstance());
                     methods.deleteAd(((MainActivity) getActivity()).userInfo, adData, new onDeleteListener() {
@@ -124,40 +83,70 @@ public class frag_ViewMyAd extends Fragment {
                         }
                     });
                 }
-            });
+                return true;
+            }
+        });
+        super.onResume();
+    }
+
+    @Nullable
+    @Override
+    public View onCreateView(LayoutInflater inflater, @Nullable ViewGroup container, Bundle savedInstanceState) {
+
+        view = inflater.inflate(R.layout.frag_viewmyad, container, false);
+
+        context = view.getContext();
+
+        initVariables();
+
+        setUpViews();
+        return view;
+    }
+
+    private void initVariables() {
+
+        tv_Title = (TextView) view.findViewById(R.id.Ad_tvTitle);
+        tv_Price = (TextView) view.findViewById(R.id.Ad_tvPrice);
+        tv_Description = (TextView) view.findViewById(R.id.Ad_tvDescription);
+        imgRecyclerView = (RecyclerView) view.findViewById(R.id.Ad_imgRecycler);
+    }
+
+    void setUpViews() {
+
+        adData = getArguments().getParcelable("adData");
+
+        if(adData!=null) {
+            Log.i(TAG,"AdData not null");
+
+            if (adData.getPrice() == 0)
+                tv_Price.setText(getString(R.string.free));
+            else
+                tv_Price.setText(String.format(getString(R.string.currency), adData.getPrice()));
+
+            tv_Title.setText(adData.getTitle());
+            tv_Description.setText(adData.getDescription());
+
+            setUpImgRecycler();
         }
         else
             Log.i("frag_ViewAd YOYO","no adDATA");
     }
 
     void setUpImgRecycler() {
-        Log.i(TAG,"setUpImgRecycler");
         main = ((frag_myAds)(getFragmentManager().findFragmentByTag(MY_ADS_TAG))).current;
-        ViewAdImageAdapter adapter = new ViewAdImageAdapter(context, adData, main, cloudStorageMethods);
+        ViewAdImageAdapter adapter = new ViewAdImageAdapter(context, adData, main);
         LinearLayoutManager linearLayoutManager = new LinearLayoutManager(context, LinearLayoutManager.HORIZONTAL, false);
         imgRecyclerView.setLayoutManager(linearLayoutManager);
         imgRecyclerView.setAdapter(adapter);
     }
 
-    //endregion
-
-    @Override
-    public void onResume() {
-        ((MainActivity)getActivity()).hideIT();
-        super.onResume();
-    }
-
     @Override
     public void onStop() {
         super.onStop();
-        ((MainActivity)context).ic_Toolbar.setVisibility(View.GONE);
-        ((MainActivity)context).ic_Toolbar.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                ((frag_Main)(getFragmentManager().findFragmentByTag(MAIN_FRAG_TAG))).refresh();
-            }
-        });
+        ((MainActivity) getActivity()).toolbar.getMenu().findItem(R.id.toolbar_delete).setVisible(false);
     }
+
+    //endregion
 
 }
